# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FV3M",
        "contact": {
            "email": "mbel002@mbel.co.jp"
        },
        "license": {
            "name": "MBEL",
            "url": "http://www.mbel.co.jp/"
        },
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://3.112.130.195:8080/api/v1",
            "description": "Staging API Server"
        },
        {
            "url": "http://fv3m.jp:8000/api/v1",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/customers/{id}": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Get customer By Id",
                "description": "Get Individual customer data according to customer-id",
                "operationId": "App\\Http\\Controllers\\CustomerController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "returns customer data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": ""
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Customers"
                ],
                "summary": "Update customer",
                "description": "updates customer data",
                "operationId": "App\\Http\\Controllers\\CustomerController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "mithun"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "new customer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "returns updated customer data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": ""
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Customers"
                ],
                "summary": "Delete customer",
                "description": "delete customer data",
                "operationId": "App\\Http\\Controllers\\CustomerController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/devices": {
            "get": {
                "tags": [
                    "Devices"
                ],
                "summary": "Get list of devices",
                "operationId": "App\\Http\\Controllers\\DeviceController::index",
                "parameters": [
                    {
                        "name": "perPage",
                        "in": "path",
                        "required": false
                    },
                    {
                        "name": "page",
                        "in": "path",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "returns list of devices with pagination .",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": ""
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied!."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Devices"
                ],
                "summary": "Store new device",
                "description": "Returns device data",
                "operationId": "App\\Http\\Controllers\\DeviceController::store",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "forklift"
                                    },
                                    "lat": {
                                        "type": "string",
                                        "example": "35.26"
                                    },
                                    "lng": {
                                        "type": "string",
                                        "example": "176.2"
                                    },
                                    "customer_id": {
                                        "type": "int",
                                        "example": "1"
                                    },
                                    "status_id": {
                                        "type": "int",
                                        "example": "1"
                                    },
                                    "imageUrls": {
                                        "type": "array",
                                        "items": {
                                            "type": "array",
                                            "items": {}
                                        },
                                        "example": "https://5.imimg.com/data5/AL/CC/MY-19161367/counterbalanced-forklift-250x250.png"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "returns stored device data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": ""
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied!"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/devices/{id}": {
            "get": {
                "tags": [
                    "Devices"
                ],
                "summary": "Get device By Id",
                "description": "Get Individual device data according to device-id",
                "operationId": "App\\Http\\Controllers\\DeviceController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "returns device data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": ""
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Devices"
                ],
                "summary": "Update device",
                "description": "updates device data",
                "operationId": "App\\Http\\Controllers\\DeviceController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "forklift"
                                    },
                                    "lat": {
                                        "type": "string",
                                        "example": "35.26"
                                    },
                                    "lng": {
                                        "type": "string",
                                        "example": "176.2"
                                    },
                                    "customer_id": {
                                        "type": "int",
                                        "example": "1"
                                    },
                                    "status_id": {
                                        "type": "int",
                                        "example": "1"
                                    },
                                    "imageUrls": {
                                        "type": "array",
                                        "items": {
                                            "type": "array",
                                            "items": {}
                                        },
                                        "example": "https://5.imimg.com/data5/AL/CC/MY-19161367/counterbalanced-forklift-250x250.png"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "returns updated devices data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": ""
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Devices"
                ],
                "summary": "Delete device",
                "description": "delete device data",
                "operationId": "App\\Http\\Controllers\\DeviceController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/devices/{deviceId}/maintenances": {
            "get": {
                "tags": [
                    "Devices"
                ],
                "summary": "Get maintenances based on device",
                "description": "Returns maintenances data based on device",
                "operationId": "App\\Http\\Controllers\\DeviceController::getMaintenances",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "name": "perPage",
                        "in": "path",
                        "required": false
                    },
                    {
                        "name": "page",
                        "in": "path",
                        "required": false
                    }
                ],
                "responses": {
                    "201": {
                        "description": "returns maintenances data based on device",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": ""
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied!"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/maintenances/{id}": {
            "get": {
                "tags": [
                    "Maintenances"
                ],
                "summary": "Get maintenance By Id",
                "description": "Get Individual maintenance data according to maintenance-id",
                "operationId": "App\\Http\\Controllers\\MaintenanceController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "returns maintenance data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": ""
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Maintenances"
                ],
                "summary": "Update maintenance",
                "description": "updates maintenance data",
                "operationId": "App\\Http\\Controllers\\MaintenanceController::edit",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "device_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "maintenece"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "camera repair"
                                    },
                                    "lat": {
                                        "type": "string",
                                        "example": "35.26"
                                    },
                                    "lng": {
                                        "type": "string",
                                        "example": "176.2"
                                    },
                                    "device_id": {
                                        "type": "int",
                                        "example": "1"
                                    },
                                    "imageUrls": {
                                        "type": "string",
                                        "example": "https://5.imimg.com/data5/AL/CC/MY-19161367/counterbalanced-forklift-250x250.png"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "returns updated maintenance data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": ""
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Maintenances"
                ],
                "summary": "Delete maintenance",
                "description": "delete maintenance data",
                "operationId": "App\\Http\\Controllers\\MaintenanceController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "A short description for security scheme",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        []
    ]
}